requirements.txt
Flask==2.3.3
requests==2.31.0
pyserial==3.5
python-escpos==3.0
flask-cors==4.0.0
python-dotenv==1.0.0
gunicorn==21.2.0
psutil==5.9.6